display count and storage location
displays count and location in inventory if item is owned
